19 CarController made in partial class for divide on version 1 & 2 its methods.
